requirements.txt
black==25.1.0
click==8.2.1
greenlet==3.2.3
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.53.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
typing_extensions==4.14.1
